print raw $BASH_COMMAND on error
in case of error, BASH_COMMAND contains
the command that throw the error (catched by
the trap).
As the command is with variables (as example:
$TORBIN" -f "$TORCONFIGFILE"), I was evaluating
it to show the real command to the user.
this is not always possible because if the command
contains a &> /dev/null this will be evaluated to
showing an empty string to user.
This restore the raw command plus a message if -v
flag was not specified.